Move and rename test helper for reuse
