+++ src/test/java/org/iti/sqlSchemaComparison/frontends/database/SqliteSchemaFrontendTest.java
@@ -28,6 +28,7 @@ import static org.junit.Assert.assertTrue;
 import java.nio.file.InvalidPathException;
 import java.util.Map.Entry;
 
+import org.iti.sqlSchemaComparison.TestHelper;
 import org.iti.sqlSchemaComparison.SchemaModification;
 import org.iti.sqlSchemaComparison.SqlSchemaColumnComparisonResult;
 import org.iti.sqlSchemaComparison.SqlSchemaComparer;
@@ -84,7 +85,7 @@ public class SqliteSchemaFrontendTest {
 		assertNotNull(schema);
 		assertEquals(8, SqlElementFactory.getSqlElementsOfType(SqlTableVertex.class, schema.vertexSet()).size());
 		assertEquals(31, SqlElementFactory.getSqlElementsOfType(SqlColumnVertex.class, schema.vertexSet()).size());
-		assertEquals(7, DatabaseFrontendTestHelper.getColumnWithConstraint(schema, IColumnConstraint.ConstraintType.PRIMARY_KEY).size());
+		assertEquals(7, TestHelper.getColumnWithConstraint(schema, IColumnConstraint.ConstraintType.PRIMARY_KEY).size());
 
 	}
 
@@ -171,7 +172,7 @@ public class SqliteSchemaFrontendTest {
 		assertEquals(31, SqlElementFactory.getSqlElementsOfType(SqlColumnVertex.class, schema1.vertexSet()).size());
 		assertEquals(31, SqlElementFactory.getSqlElementsOfType(SqlColumnVertex.class, schema2.vertexSet()).size());
 
-		Entry<ISqlElement, SchemaModification> entry = DatabaseFrontendTestHelper.getModificationOfType(result, SchemaModification.MOVE_COLUMN);
+		Entry<ISqlElement, SchemaModification> entry = TestHelper.getModificationOfType(result, SchemaModification.MOVE_COLUMN);
 
 		assertNotNull(entry);
 		assertEquals(SchemaModification.MOVE_COLUMN, entry.getValue());
@@ -190,7 +191,7 @@ public class SqliteSchemaFrontendTest {
 		assertEquals(31, SqlElementFactory.getSqlElementsOfType(SqlColumnVertex.class, schema1.vertexSet()).size());
 		assertEquals(31, SqlElementFactory.getSqlElementsOfType(SqlColumnVertex.class, schema2.vertexSet()).size());
 
-		Entry<ISqlElement, SchemaModification> entry = DatabaseFrontendTestHelper.getModificationOfType(result, SchemaModification.RENAME_COLUMN);
+		Entry<ISqlElement, SchemaModification> entry = TestHelper.getModificationOfType(result, SchemaModification.RENAME_COLUMN);
 
 		assertNotNull(entry);
 		assertEquals(SchemaModification.RENAME_COLUMN, entry.getValue());
